refactor(plugin): extract audio_clock.rs (#1079)
Moves the 3 host_acv_* callbacks + OnTickBridge struct +
HOST_AUDIO_CLOCK_VTABLE + host_audio_clock_vtable() accessor + the
two test modules (null-handle guards + tier-1 wire-format) out of
host_services/mod.rs into a sibling submodule. mod.rs re-exports
HOST_AUDIO_CLOCK_VTABLE and host_audio_clock_vtable.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
